Return task to reducer from response. Customize Task and Item  modle validations. Implement updateTask thunk action. taskUpdated Redux action.
