+++ client/src/actions/projectActions.js
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import { ADD_PROJECT, SET_PROJECTS, PROJECT_FETCHED, PROJECT_UPDATED, PROJECT_DELETED, ADD_TASK } from './types'
+import { ADD_PROJECT, SET_PROJECTS, PROJECT_FETCHED, PROJECT_UPDATED, PROJECT_DELETED, ADD_TASK, TASK_UPDATED, TASK_DELETED } from './types'
 
 export function addProject(project) {
   return {
@@ -40,6 +40,13 @@ export function addTask(task) {
   return {
     type: ADD_TASK,
     task  
+  }
+}
+
+export function taskUpdated(task) {
+  return {
+    type: TASK_UPDATED,
+    task
   }
 }
 
@@ -86,12 +93,17 @@ export function deleteProject(id) {
 export function createTask(task) {
   return dispatch => {
     return axios.post('/api/tasks', task).then(res => {
-      dispatch(addTask(task))
+      dispatch(addTask(res.data.result))
     })
   }
 }
 
 export function updateTask(task) {
+  return dispatch => {
+    return axios.put(`/api/tasks/${task._id}`, task).then(res => {
+      dispatch(taskUpdated(res.data.result))
+    })
+  }
   
 }
