+++ client/src/components/Projects/Show.js
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom'
 import classnames from 'classnames'
 import { fetchProject, deleteProject } from '../../actions/projectActions'
 
-import TaskIndex from './Tasks/Page'
+import TaskForm from './Tasks/Form'
 
 // Localization 
 import T from 'i18n-react'
@@ -74,7 +74,7 @@ class Show extends Component {
 
             <h3 className="ui header">{T.translate("projects.tasks.header")}</h3>
 
-            { this.state.tasks && <TaskIndex tasks={this.state.tasks} /> }
+            { this.state.tasks && <TaskForm creator={this.state._id} tasks={this.state.tasks} /> }
             
             <div className="ui divider"></div>
