+++ server/src/models/Item.js
@@ -3,11 +3,11 @@ import Sale from './Sale'
 
 let itemSchema = new mongoose.Schema({
   _creator: { type: mongoose.Schema.Types.ObjectId, ref: "Sale" },
-  name: { type: String, required: true },
-  unit: { type: String, required: true },
-  quantity: { type: Number, required: true },
-  price: { type: Number, required: true },
-  vat: { type: Number, min: 1, max: 100, required: true },
+  name: { type: String, required: [true, "Name is required."] },
+  unit: { type: String, required: [true, "Unit is required."] },
+  quantity: { type: Number, required: [true, "Quantity is required."] },
+  price: { type: Number, required: [true, "Price is required."] },
+  vat: { type: Number, min: 1, max: 100, required: [true, "Vat is required."] },
 
   created_at: Date,
   updated_at: Date
